requirements.txt
aiogram>2.21
asyncpg>0.26.0
Pillow>9.2.0
pytube~=15.0.0
SQLAlchemy>1.4.39
python-dotenv~=1.0.0